PRE18

+++ src/main/java/io/github/moulberry/notenoughupdates/dungeons/GuiDungeonMapEditor.java
@@ -523,7 +523,7 @@ public class GuiDungeonMapEditor extends GuiScreen {
 
             try {
                 blurField.setCustomBorderColour(-1);
-                NotEnoughUpdates.INSTANCE.config.dungeonMap.dmBackgroundBlur = (int)Float.parseFloat(blurField.getText());
+                NotEnoughUpdates.INSTANCE.config.dungeonMap.dmBackgroundBlur = Float.parseFloat(blurField.getText());
             } catch(Exception e) {
                 blurField.setCustomBorderColour(Color.RED.getRGB());
             }
